moved createStation to stationfactory
